Change to outgoing-webhook parsing
I still am not a huge fan of how I'm handeling this, but this change
assumes that the incominging webhook is a single character.  I kinda
liked how the syntax turned out.  In my outgoing webhook configuration,
I have my Trigger Word set to "!", which makes the commands look decent
in chat:
  !ping
  !wiki [Arguments]

So, this change is minimal, but a significant difference in how the bot
interprets outgoing webhooks.